Move operation helper functions `api/common.rs`
