--- crates/moss-collection/src/api/batch_delete_entry.rs
@@ -1,1 +1,0 @@
-
