+++ crates/moss-collection/src/api/create_entry.rs
@@ -1,18 +1,8 @@
 use moss_common::api::OperationResult;
-use validator::Validate;
 
 use crate::{
     collection::Collection,
-    models::{
-        operations::{
-            CreateDirEntryInput, CreateEntryInput, CreateEntryOutput, CreateItemEntryInput,
-        },
-        primitives::EntryId,
-        types::configuration::{
-            CompositeDirConfigurationModel, CompositeItemConfigurationModel, ConfigurationMetadata,
-        },
-    },
-    services::worktree_service::{EntryMetadata, WorktreeService},
+    models::operations::{CreateEntryInput, CreateEntryOutput},
 };
 
 impl Collection {
@@ -24,65 +14,5 @@ impl Collection {
             CreateEntryInput::Item(input) => self.create_item_entry(input).await,
             CreateEntryInput::Dir(input) => self.create_dir_entry(input).await,
         }
-    }
-
-    pub(crate) async fn create_dir_entry(
-        &self,
-        input: CreateDirEntryInput,
-    ) -> OperationResult<CreateEntryOutput> {
-        input.validate()?;
-
-        let worktree_service = self.service::<WorktreeService>();
-
-        let id = EntryId::new();
-        let model = CompositeDirConfigurationModel {
-            metadata: ConfigurationMetadata { id: id.to_string() },
-            inner: input.configuration,
-        };
-
-        worktree_service
-            .create_dir_entry(
-                &id,
-                &input.name,
-                input.path,
-                model.into(),
-                EntryMetadata {
-                    order: input.order,
-                    expanded: true, // Directories are automatically marked as expanded
-                },
-            )
-            .await?;
-
-        Ok(CreateEntryOutput { id: id })
-    }
-
-    pub(crate) async fn create_item_entry(
-        &self,
-        input: CreateItemEntryInput,
-    ) -> OperationResult<CreateEntryOutput> {
-        input.validate()?;
-
-        let worktree_service = self.service::<WorktreeService>();
-
-        let id = EntryId::new();
-        let model = CompositeItemConfigurationModel {
-            metadata: ConfigurationMetadata { id: id.to_string() },
-            inner: input.configuration,
-        };
-
-        worktree_service
-            .create_item_entry(
-                &id,
-                &input.name,
-                input.path,
-                model.into(),
-                EntryMetadata {
-                    order: input.order,
-                    expanded: false,
-                },
-            )
-            .await?;
-
-        Ok(CreateEntryOutput { id: id })
     }
 }
